Refactored code to not use docker
