+++ integrations/phishing_analyzers/requirements.txt
@@ -1,4 +1,3 @@
 Flask-Shell2HTTP-fork==1.9.2
 gunicorn==23.0.0
 seleniumbase[selenium-wire]==4.30.8
-lxml==5.3.0
